add interaction routes
