+++ src/blueprints/events/__init__.py
@@ -1,5 +1,38 @@
-from sanic import Blueprint
-from src.blueprints.events.amethyst import amethyst_blueprint
+from sanic import Blueprint, Request
+import sanic
 
-event_blueprints = [amethyst_blueprint]
-event_blueprint_group = Blueprint.group(url_prefix='/events', *event_blueprints)
+from src.database import Database
+
+from sanic_ext import openapi
+
+from src.models import events, user
+
+events_blueprint = Blueprint('events', '/events')
+
+
+@events_blueprint.route('/connection', methods=['POST'])
+@openapi.body(content={'application/json': events.ConnectionCreateModel.model_json_schema()})
+@openapi.response(status=201, description='Success')
+async def connect_event(request: Request, db: Database):
+    model = events.ConnectionCreateModel(**request.json)
+    user_playtime = await user.PlaytimeSummary.fetch(db, model.thorny_id)
+
+    if model.type == 'connect' and user_playtime.session is None:
+        await events.ConnectionModel.new(db, model)
+    elif model.type == 'disconnect' and user_playtime.session:
+        await events.ConnectionModel.new(db, model)
+    else:
+        return sanic.HTTPResponse(status=400)
+
+    return sanic.HTTPResponse(status=201)
+
+
+@events_blueprint.route('/interaction', methods=['POST'])
+@openapi.body(content={'application/json': events.InteractionCreateModel.model_json_schema()})
+@openapi.response(status=201, description='Success')
+async def interaction_event(request: Request, db: Database):
+    model = events.InteractionCreateModel(**request.json)
+
+    await events.InteractionModel.new(db, model)
+
+    return sanic.HTTPResponse(status=201)
